safari-600.7-branch/LayoutTests/media/track/audio/audio-track-mkv-vorbis-language.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../../media-file.js></script>
        <script src=../../video-test.js></script>
        <script src=../../in-band-tracks.js></script>
    </head>
    <body onload="testAttribute('../../content/two-audio-and-video-tracks.mkv', 'audio', 'language', ['la', 'en'])">
        <video controls></video>
        <p>Check audio tracks' language attributes.</p>
    </body>
</html>
