LayoutTests/media/track/audio/audio-track-mkv-vorbis-language-expected.txt


Check audio tracks' language attributes.

EVENT(canplaythrough)

** Check in-band kind attributes
EXPECTED (video.audioTracks.length == '2') OK
EXPECTED (video.audioTracks[0].language == 'en') OK
EXPECTED (video.audioTracks[1].language == 'la') OK

END OF TEST
